fix: move tests to another file for AI purposes
